requirements.txt
pip>=10.0.1
falcon>=1.3.0
uwsgi>=2.0.17
mysql-connector-python==8.0.6
pyjwt>=1.5.3
requests>=2.18.4
cryptography
